ImageOverlay update and pluginification, and example nb
