+++ folium/plugins/__init__.py
@@ -12,10 +12,13 @@ from .terminator import Terminator
 from .boat_marker import BoatMarker
 from .timestamped_geo_json import TimestampedGeoJson
 from .heat_map import HeatMap
+from .image_overlay import ImageOverlay
 
 __all__ = ['MarkerCluster',
            'ScrollZoomToggler',
            'Terminator',
            'BoatMarker',
            'TimestampedGeoJson',
-           'HeatMap']
+           'HeatMap',
+           'ImageOverlay',
+          ]
